requirements.txt
# Core dependencies for the project
python-dotenv>=1.0.0

# Database
psycopg2-binary>=2.9.0

# API Framework
fastapi>=0.104.0
uvicorn>=0.24.0

# ML/AI
torch>=2.0.0
transformers>=4.35.0
tokenizers>=0.14.0
pytorch-lightning>=2.6.0

# MLflow
mlflow>=2.5.0

# Data processing
pandas>=2.1.0
numpy>=1.26.0
scikit-learn>=1.3.0

# Language detection
langdetect>=1.0.9

# Utilities
pydantic>=2.0.0

